+++ cmd/neofs-node/transport.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 
 	objectGRPC "github.com/nspcc-dev/neofs-api-go/v2/object/grpc"
-	"github.com/nspcc-dev/neofs-api-go/v2/refs"
 	rawclient "github.com/nspcc-dev/neofs-api-go/v2/rpc/client"
 	"github.com/nspcc-dev/neofs-api-go/v2/rpc/common"
 	"github.com/nspcc-dev/neofs-api-go/v2/rpc/grpc"
@@ -13,7 +12,6 @@ import (
 	"github.com/nspcc-dev/neofs-api-go/v2/status"
 	coreclient "github.com/nspcc-dev/neofs-node/pkg/core/client"
 	apistatus "github.com/nspcc-dev/neofs-sdk-go/client/status"
-	neofscrypto "github.com/nspcc-dev/neofs-sdk-go/crypto"
 )
 
 type transport struct {
@@ -22,7 +20,7 @@ type transport struct {
 
 // SendReplicationRequestToNode connects to described node and sends prepared
 // replication request message to it.
-func (x *transport) SendReplicationRequestToNode(ctx context.Context, req []byte, node coreclient.NodeInfo) (*neofscrypto.Signature, error) {
+func (x *transport) SendReplicationRequestToNode(ctx context.Context, req []byte, node coreclient.NodeInfo) ([]byte, error) {
 	c, err := x.clients.Get(node)
 	if err != nil {
 		return nil, fmt.Errorf("connect to remote node: %w", err)
@@ -40,12 +38,12 @@ func (x *transport) SendReplicationRequestToNode(ctx context.Context, req []byte
 		}
 		return resp.err
 	})
-	return resp.sig, err
+	return resp.sigs, err
 }
 
 type replicateResponse struct {
-	sig *neofscrypto.Signature
-	err error
+	sigs []byte
+	err  error
 }
 
 func (x replicateResponse) ToGRPCMessage() grpc.Message { return new(objectGRPC.ReplicateResponse) }
@@ -70,22 +68,7 @@ func (x *replicateResponse) FromGRPCMessage(gm grpc.Message) error {
 		return nil
 	}
 
-	sig := m.GetObjectSignature()
-	if sig == nil {
-		return nil
-	}
-
-	sigV2 := new(refs.Signature)
-	err := sigV2.Unmarshal(sig)
-	if err != nil {
-		return fmt.Errorf("decoding signature from proto message: %w", err)
-	}
-
-	x.sig = new(neofscrypto.Signature)
-	err = x.sig.ReadFromV2(*sigV2)
-	if err != nil {
-		return fmt.Errorf("invalid signature: %w", err)
-	}
+	x.sigs = m.GetObjectSignature()
 
 	return nil
 }