+++ pkg/innerring/processors/container/process_container.go
@@ -26,7 +26,10 @@ type putEvent interface {
 type putContainerContext struct {
 	e putEvent
 
-	d containerSDK.Domain
+	// must be filled when verifying raw data from e
+	cID cid.ID
+	cnr containerSDK.Container
+	d   containerSDK.Domain
 }
 
 // Process a new container from the user by checking the container sanity
@@ -55,15 +58,15 @@ func (cp *Processor) processContainerPut(put putEvent) {
 
 func (cp *Processor) checkPutContainer(ctx *putContainerContext) error {
 	binCnr := ctx.e.Container()
-	var cnr containerSDK.Container
+	ctx.cID = cid.NewFromMarshalledContainer(binCnr)
 
-	err := cnr.Unmarshal(binCnr)
+	err := ctx.cnr.Unmarshal(binCnr)
 	if err != nil {
 		return fmt.Errorf("invalid binary container: %w", err)
 	}
 
 	err = cp.verifySignature(signatureVerificationData{
-		ownerContainer:  cnr.Owner(),
+		ownerContainer:  ctx.cnr.Owner(),
 		verb:            session.VerbContainerPut,
 		binTokenSession: ctx.e.SessionToken(),
 		binPublicKey:    ctx.e.PublicKey(),
@@ -75,13 +78,13 @@ func (cp *Processor) checkPutContainer(ctx *putContainerContext) error {
 	}
 
 	// check homomorphic hashing setting
-	err = checkHomomorphicHashing(cp.netState, cnr)
+	err = checkHomomorphicHashing(cp.netState, ctx.cnr)
 	if err != nil {
 		return fmt.Errorf("incorrect homomorphic hashing setting: %w", err)
 	}
 
 	// check native name and zone
-	err = checkNNS(ctx, cnr)
+	err = checkNNS(ctx, ctx.cnr)
 	if err != nil {
 		return fmt.Errorf("NNS: %w", err)
 	}
@@ -94,22 +97,38 @@ func (cp *Processor) approvePutContainer(ctx *putContainerContext) {
 
 	var err error
 
-	prm := cntClient.PutPrm{}
-
-	prm.SetContainer(e.Container())
-	prm.SetKey(e.PublicKey())
-	prm.SetSignature(e.Signature())
-	prm.SetToken(e.SessionToken())
-	prm.SetName(ctx.d.Name())
-	prm.SetZone(ctx.d.Zone())
-
 	nr := e.NotaryRequest()
-	err = cp.cnrClient.Morph().NotarySignAndInvokeTX(nr.MainTransaction)
+	err = cp.cnrClient.Morph().NotarySignAndInvokeTX(nr.MainTransaction, true)
 
 	if err != nil {
 		cp.log.Error("could not approve put container",
 			zap.Error(err),
 		)
+		return
+	}
+
+	nm, err := cp.netState.NetMap()
+	if err != nil {
+		cp.log.Error("could not get netmap for Container contract update", zap.Stringer("cid", ctx.cID), zap.Error(err))
+		return
+	}
+
+	policy := ctx.cnr.PlacementPolicy()
+	vectors, err := nm.ContainerNodes(policy, ctx.cID)
+	if err != nil {
+		cp.log.Error("could not build placement for Container contract update", zap.Stringer("cid", ctx.cID), zap.Error(err))
+		return
+	}
+
+	replicas := make([]uint32, 0, policy.NumberOfReplicas())
+	for i := range vectors {
+		replicas = append(replicas, policy.ReplicaNumberByIndex(i))
+	}
+
+	err = cp.cnrClient.UpdateContainerPlacement(ctx.cID, vectors, replicas)
+	if err != nil {
+		cp.log.Error("could not update Container contract", zap.Stringer("cid", ctx.cID), zap.Error(err))
+		return
 	}
 }
 
@@ -175,7 +194,7 @@ func (cp *Processor) approveDeleteContainer(e *containerEvent.Delete) {
 	prm.SetToken(e.SessionToken())
 
 	nr := e.NotaryRequest()
-	err = cp.cnrClient.Morph().NotarySignAndInvokeTX(nr.MainTransaction)
+	err = cp.cnrClient.Morph().NotarySignAndInvokeTX(nr.MainTransaction, false)
 
 	if err != nil {
 		cp.log.Error("could not approve delete container",