+++ pkg/morph/client/container/client.go
@@ -39,6 +39,10 @@ const (
 
 	// putNamedMethod is method name for container put with an alias. It is exported to provide custom fee.
 	putNamedMethod = "putNamed"
+
+	addNextEpochNodes         = "addNextEpochNodes"
+	commitContainerListUpdate = "commitContainerListUpdate"
+	submitObjectPutMethod     = "submitObjectPut"
 )
 
 var (