+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog for NeoFS Node
 ## [Unreleased]
 
 ### Added
+- Initial support for meta-on-chain for objects (#2877)
 
 ### Fixed